1. Move message service related code to a new package 'server/src/main/java/com/broadcom/lsp/cobol/core/messages' 2. Reformat code based on google-java-formatter plugin
