+++ server/src/main/java/com/broadcom/lsp/cobol/core/messages/ExternalizedMessage.java
@@ -12,7 +12,7 @@
  *    Broadcom, Inc. - initial API and implementation
  *
  */
-package com.broadcom.lsp.cobol.core.strategy;
+package com.broadcom.lsp.cobol.core.messages;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,18 +22,18 @@ import java.util.MissingFormatArgumentException;
 
 @AllArgsConstructor
 @Builder
-public class ExternalizeMessage {
-    private final MessageTemplate messageTemplate;
-    private final Object[] parameters;
-    private final String key;
+public class ExternalizedMessage {
+  private final MessageTemplate messageTemplate;
+  private final Object[] parameters;
+  private final String key;
 
-    /***
-     *
-     * @return a formatted text for a {@link ExternalizeMessage} object.
-     * @throws MissingFormatArgumentException   when the amount of parameters is less than required.
-     * @throws IllegalFormatConversionException when format is invalid.
-     */
-    public String getFormattedMessage() {
-        return String.format(messageTemplate.getMessage(), parameters);
-    }
+  /***
+   *
+   * @return a formatted text for a {@link ExternalizedMessage} object.
+   * @throws MissingFormatArgumentException   when the amount of parameters is less than required.
+   * @throws IllegalFormatConversionException when format is invalid.
+   */
+  public String getFormattedMessage() {
+    return String.format(messageTemplate.getMessage(), parameters);
+  }
 }
