+++ server/src/main/java/com/broadcom/lsp/cobol/core/messages/MessageService.java
@@ -12,18 +12,19 @@
  *    Broadcom, Inc. - initial API and implementation
  *
  */
-package com.broadcom.lsp.cobol.core.strategy;
+package com.broadcom.lsp.cobol.core.messages;
 
 import java.util.List;
 
 public interface MessageService {
 
-    String INVALID_KEY = "invalid-key";
-    String INVALID_KEY_S_SUPPLIED_PLEASE_CHECK_THE_EXTERNALIZED_MESSAGE_FILES = "Invalid key '%s' supplied. Please check the externalized message files.";
+  String INVALID_KEY = "invalid-key";
+  String INVALID_KEY_S_SUPPLIED_PLEASE_CHECK_THE_EXTERNALIZED_MESSAGE_FILES =
+      "Invalid key '%s' supplied. Please check the externalized message files.";
 
-    ExternalizeMessage getMessage(String key, Object... parameters);
+  ExternalizedMessage getMessage(String key, Object... parameters);
 
-    void loadMessages(String filePath);
+  void loadMessages(String filePath);
 
-    void addMessageTemplates(List<MessageTemplate> messageTemplateList);
+  void addMessageTemplates(List<MessageTemplate> messageTemplateList);
 }
