+++ server/src/main/java/com/broadcom/lsp/cobol/core/messages/MessageTemplate.java
@@ -12,10 +12,20 @@
  *    Broadcom, Inc. - initial API and implementation
  *
  */
-package com.broadcom.lsp.cobol.core.strategy;
+package com.broadcom.lsp.cobol.core.messages;
 
-public class MessageTemplateLoadException extends RuntimeException {
-    public MessageTemplateLoadException(String format) {
-        super(format);
-    }
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageTemplate {
+  private String key;
+  private String message;
+
+  public boolean isValidKey() {
+    return !key.equals(MessageService.INVALID_KEY);
+  }
 }
